Make layer alpha configurable and generalize Rgb/Grayscale layer handling (#3670)
* combine abstract_plane_material_factory and plane_material_factory

* add alpha setting to each color layer, allow to have rgb and color layers in the same dataset, design tweaks for layer settings

* fix flow

* update changelog

* apply PR feedback

* make layer name bold in settings

* pretty
