+++ app/assets/javascripts/oxalis/geometries/materials/edge_shader.js
@@ -3,7 +3,7 @@
 import * as THREE from "three";
 
 import { COLOR_TEXTURE_WIDTH_FIXED } from "oxalis/geometries/materials/node_shader";
-import type { Uniforms } from "oxalis/geometries/materials/abstract_plane_material_factory";
+import type { Uniforms } from "oxalis/geometries/materials/plane_material_factory";
 
 class EdgeShader {
   material: THREE.RawShaderMaterial;
