+++ app/assets/javascripts/oxalis/geometries/materials/plane_material_factory.js
@@ -1,11 +1,8 @@
-/**
- * plane_material_factory.js
- * @flow
- */
-
+// @flow
 import * as THREE from "three";
 import _ from "lodash";
 
+import app from "app";
 import {
   ModeValues,
   type OrthoView,
@@ -31,29 +28,67 @@ import {
 import { getPackingDegree } from "oxalis/model/bucket_data_handling/data_rendering_logic";
 import { getViewportScale } from "oxalis/model/accessors/view_mode_accessor";
 import { listenToStoreProperty } from "oxalis/model/helpers/listener_helpers";
-import AbstractPlaneMaterialFactory, {
-  type ShaderMaterialOptions,
-  sanitizeName,
-} from "oxalis/geometries/materials/abstract_plane_material_factory";
 import Model from "oxalis/model";
 import Store, { type DatasetLayerConfiguration } from "oxalis/store";
 import * as Utils from "libs/utils";
 import getMainFragmentShader from "oxalis/shaders/main_data_fragment.glsl";
+import shaderEditor from "oxalis/model/helpers/shader_editor";
+
+type ShaderMaterialOptions = {
+  polygonOffset?: boolean,
+  polygonOffsetFactor?: number,
+  polygonOffsetUnits?: number,
+};
+
+export type Uniforms = {
+  [key: string]: {
+    type: "b" | "f" | "i" | "t" | "v2" | "v3" | "v4" | "tv",
+    value: any,
+  },
+};
 
 const DEFAULT_COLOR = new THREE.Vector3([255, 255, 255]);
+
+function sanitizeName(name: ?string): string {
+  // Make sure name starts with a letter and contains no "-" characters
+  if (name == null) {
+    return "";
+  }
+  return `layer_${name.replace(/-/g, "_")}`;
+}
 
 function getColorLayerNames() {
   return getColorLayers(Store.getState().dataset).map(layer => sanitizeName(layer.name));
 }
 
-class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
+function getIsRgbLayerLookup(): { [string]: boolean } {
+  const { dataset } = Store.getState();
+  const colorLayers = getColorLayers(dataset);
+  // keyBy the sanitized layer name as the lookup will happen in the shader using the sanitized layer name
+  const colorLayersObject = _.keyBy(colorLayers, layer => sanitizeName(layer.name));
+  return _.mapValues(colorLayersObject, layer => isRgb(dataset, layer.name));
+}
+
+class PlaneMaterialFactory {
   planeID: OrthoView;
   isOrthogonal: boolean;
+  material: THREE.ShaderMaterial;
+  uniforms: Uniforms;
+  attributes: Object;
+  shaderId: number;
+  storePropertyUnsubscribers: Array<() => void> = [];
 
   constructor(planeID: OrthoView, isOrthogonal: boolean, shaderId: number) {
-    super(shaderId);
     this.planeID = planeID;
     this.isOrthogonal = isOrthogonal;
+    this.shaderId = shaderId;
+  }
+
+  setup() {
+    this.setupUniforms();
+    this.makeMaterial();
+    this.attachTextures();
+    return this;
   }
 
   stopListening() {
@@ -61,9 +96,7 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
   }
 
   setupUniforms(): void {
-    super.setupUniforms();
-
-    this.uniforms = _.extend(this.uniforms, {
+    this.uniforms = {
       alpha: {
         type: "f",
         value: 0,
@@ -164,12 +197,31 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
         type: "v3",
         value: new THREE.Vector3(0, 0, 0),
       },
-    });
+    };
 
     for (const dataLayer of Model.getAllLayers()) {
       this.uniforms[sanitizeName(`${dataLayer.name}_maxZoomStep`)] = {
         type: "f",
         value: dataLayer.cube.MAX_ZOOM_STEP,
+      };
+    }
+
+    for (const name of getColorLayerNames()) {
+      this.uniforms[`${name}_alpha`] = {
+        type: "f",
+        value: 1,
+      };
+      this.uniforms[`${name}_color`] = {
+        type: "v3",
+        value: DEFAULT_COLOR,
+      };
+      this.uniforms[`${name}_brightness`] = {
+        type: "f",
+        value: 1.0,
+      };
+      this.uniforms[`${name}_contrast`] = {
+        type: "f",
+        value: 1.0,
       };
     }
   }
@@ -221,22 +273,18 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
         value: mappingColorTexture,
       };
     }
-
-    // Add weight/color uniforms
-    for (const name of getColorLayerNames()) {
-      this.uniforms[`${name}_weight`] = {
-        type: "f",
-        value: 1,
-      };
-      this.uniforms[`${name}_color`] = {
-        type: "v3",
-        value: DEFAULT_COLOR,
-      };
-    }
   }
 
   makeMaterial(options?: ShaderMaterialOptions): void {
-    super.makeMaterial(options);
+    this.material = new THREE.ShaderMaterial(
+      _.extend(options, {
+        uniforms: this.uniforms,
+        vertexShader: this.getVertexShader(),
+        fragmentShader: this.getFragmentShader(),
+      }),
+    );
+
+    shaderEditor.addMaterial(this.shaderId, this.material);
 
     this.material.setGlobalPosition = ([x, y, z]) => {
       this.uniforms.globalPosition.value.set(x, y, z);
@@ -385,6 +433,24 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
       ),
     );
 
+    this.storePropertyUnsubscribers.push(
+      listenToStoreProperty(
+        state => state.datasetConfiguration.layers,
+        layerSettings => {
+          for (const colorLayer of getColorLayers(Store.getState().dataset)) {
+            const settings = layerSettings[colorLayer.name];
+            if (settings != null) {
+              const name = sanitizeName(colorLayer.name);
+              this.updateUniformsForLayer(settings, name);
+            }
+          }
+          // TODO: Needed?
+          app.vent.trigger("rerender");
+        },
+        true,
+      ),
+    );
+
     const hasSegmentation = Model.getSegmentationLayer() != null;
     if (hasSegmentation) {
       this.storePropertyUnsubscribers.push(
@@ -466,7 +532,9 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
   }
 
   updateUniformsForLayer(settings: DatasetLayerConfiguration, name: string): void {
-    super.updateUniformsForLayer(settings, name);
+    this.uniforms[`${name}_brightness`].value = settings.brightness / 255;
+    this.uniforms[`${name}_contrast`].value = settings.contrast;
+    this.uniforms[`${name}_alpha`].value = settings.alpha / 100;
 
     if (settings.color != null) {
       const color = this.convertColor(settings.color);
@@ -474,8 +542,13 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
     }
   }
 
+  getMaterial(): THREE.ShaderMaterial {
+    return this.material;
+  }
+
   getFragmentShader(): string {
     const colorLayerNames = getColorLayerNames();
+    const isRgbLayerLookup = getIsRgbLayerLookup();
     const segmentationLayer = Model.getSegmentationLayer();
     const segmentationName = sanitizeName(segmentationLayer ? segmentationLayer.name : "");
     const { dataset } = Store.getState();
@@ -489,10 +562,10 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
 
     const code = getMainFragmentShader({
       colorLayerNames,
+      isRgbLayerLookup,
       hasSegmentation,
       segmentationName,
       segmentationPackingDegree,
-      isRgb: Model.dataLayers.color && isRgb(dataset, Model.dataLayers.color.name),
       isMappingSupported: Model.isMappingSupported,
       dataTextureCountPerLayer: Model.maximumDataTextureCountForLayer,
       resolutions: getResolutions(dataset),
@@ -501,6 +574,24 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
     });
 
     return code;
+  }
+
+  getVertexShader(): string {
+    return `
+precision highp float;
+
+varying vec4 worldCoord;
+varying vec4 modelCoord;
+varying vec2 vUv;
+varying mat4 savedModelMatrix;
+
+void main() {
+  vUv = uv;
+  modelCoord = vec4(position, 1.0);
+  savedModelMatrix = modelMatrix;
+  worldCoord = modelMatrix * vec4(position, 1.0);
+  gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
+}`;
   }
 }
