+++ app/assets/javascripts/oxalis/store.js
@@ -208,6 +208,7 @@ export type DatasetLayerConfiguration = {|
   +color: Vector3,
   +brightness: number,
   +contrast: number,
+  +alpha: number,
 |};
 
 export type DatasetConfiguration = {|
