+++ app/assets/javascripts/oxalis/view/settings/dataset_settings_view.js
@@ -4,7 +4,7 @@
  */
 
 import type { Dispatch } from "redux";
-import { Tooltip, Collapse, Row, Col, Select, Icon } from "antd";
+import { Tooltip, Collapse, Row, Col, Select, Icon, Divider, Tag } from "antd";
 import { connect } from "react-redux";
 import * as React from "react";
 import _ from "lodash";
@@ -16,7 +16,7 @@ import {
   DropdownSetting,
   ColorSetting,
 } from "oxalis/view/settings/setting_input_views";
-import { hasSegmentation } from "oxalis/model/accessors/dataset_accessor";
+import { hasSegmentation, isRgb } from "oxalis/model/accessors/dataset_accessor";
 import {
   updateDatasetSettingAction,
   updateLayerSettingAction,
@@ -25,12 +25,14 @@ import Toast from "libs/toast";
 import * as Utils from "libs/utils";
 import constants, { type ControlMode, ControlModeEnum, type Mode } from "oxalis/constants";
 import messages, { settings } from "messages";
+import type { APIDataset } from "admin/api_flow_types";
 
 const Panel = Collapse.Panel;
 const Option = Select.Option;
 
 type DatasetSettingsProps = {
   datasetConfiguration: DatasetConfiguration,
+  dataset: APIDataset,
   onChange: (propertyName: $Keys<DatasetConfiguration>, value: any) => void,
   onChangeLayer: (
     layerName: string,
@@ -43,35 +45,57 @@ type DatasetSettingsProps = {
 };
 
 class DatasetSettings extends React.PureComponent<DatasetSettingsProps> {
-  getColorSettings = (layer: Object, layerName: string) => (
-    <div key={layerName}>
-      <Row>
-        <Col span={24}>Layer: {layerName}</Col>
-      </Row>
-      <NumberSliderSetting
-        label="Brightness"
-        min={-255}
-        max={255}
-        step={5}
-        value={layer.brightness}
-        onChange={_.partial(this.props.onChangeLayer, layerName, "brightness")}
-      />
-      <NumberSliderSetting
-        label="Contrast"
-        min={0.5}
-        max={5}
-        step={0.1}
-        value={layer.contrast}
-        onChange={_.partial(this.props.onChangeLayer, layerName, "contrast")}
-      />
-      <ColorSetting
-        label="Color"
-        value={Utils.rgbToHex(layer.color)}
-        onChange={_.partial(this.props.onChangeLayer, layerName, "color")}
-        className="ant-btn"
-      />
-    </div>
-  );
+  getColorSettings = (
+    [layerName: string, layer: DatasetLayerConfiguration],
+    layerIndex: number,
+    isLastLayer: boolean,
+  ) => {
+    const isRGB = isRgb(this.props.dataset, layerName);
+    // $FlowFixMe Object.entries returns mixed for Flow
+    const { brightness, contrast, alpha, color } = layer;
+    return (
+      <div key={layerName}>
+        <Row>
+          <Col span={24}>
+            <span style={{ fontWeight: 700 }}>{layerName}</span>
+            <Tag style={{ cursor: "default", marginLeft: 8 }} color={isRGB && "#1890ff"}>
+              {isRGB ? "24-bit" : "8-bit"} Layer
+            </Tag>
+          </Col>
+        </Row>
+        <NumberSliderSetting
+          label="Brightness"
+          min={-255}
+          max={255}
+          step={5}
+          value={brightness}
+          onChange={_.partial(this.props.onChangeLayer, layerName, "brightness")}
+        />
+        <NumberSliderSetting
+          label="Contrast"
+          min={0.5}
+          max={5}
+          step={0.1}
+          value={contrast}
+          onChange={_.partial(this.props.onChangeLayer, layerName, "contrast")}
+        />
+        <NumberSliderSetting
+          label="Opacity"
+          min={0}
+          max={100}
+          value={alpha}
+          onChange={_.partial(this.props.onChangeLayer, layerName, "alpha")}
+        />
+        <ColorSetting
+          label="Color"
+          value={Utils.rgbToHex(color)}
+          onChange={_.partial(this.props.onChangeLayer, layerName, "color")}
+          className="ant-btn"
+        />
+        {!isLastLayer && <Divider />}
+      </div>
+    );
+  };
 
   onChangeQuality = (propertyName: $Keys<DatasetConfiguration>, value: string) => {
     this.props.onChange(propertyName, parseInt(value));
@@ -115,11 +139,14 @@ class DatasetSettings extends React.PureComponent<DatasetSettingsProps> {
   }
 
   render() {
-    const colorSettings = _.map(this.props.datasetConfiguration.layers, this.getColorSettings);
+    const { layers } = this.props.datasetConfiguration;
+    const colorSettings = Object.entries(layers).map((entry, index) =>
+      this.getColorSettings(entry, index, index === _.size(layers) - 1),
+    );
 
     return (
       <Collapse defaultActiveKey={["1", "2", "3", "4"]}>
-        <Panel header="Colors" key="1">
+        <Panel header="Color Layers" key="1">
           {colorSettings}
         </Panel>
         {this.props.hasSegmentation ? this.getSegmentationPanel() : null}
@@ -167,6 +194,7 @@ const mapStateToProps = (state: OxalisState) => ({
   datasetConfiguration: state.datasetConfiguration,
   viewMode: state.temporaryConfiguration.viewMode,
   controlMode: state.temporaryConfiguration.controlMode,
+  dataset: state.dataset,
   hasSegmentation: hasSegmentation(state.dataset),
 });
