+++ app/assets/stylesheets/trace_view/_tracing_view.less
@@ -106,7 +106,7 @@
     font-size: 14px;
   }
 
-  .settings-row {
+  .margin-bottom {
     margin-bottom: 10px;
   }
